doc: mention v4.2.1 blobstore bug in sourcegraph/server (#45432)
This adds documentation to mention a v4.2.1 blobstore bug in `sourcegraph/server`.

```
failed to create bucket: operation error S3: CreateBucket, retry quota exceeded, 0 available, 5 requested
```

This only affects `sourcegraph/server`, and only appears to happen when
the data directory (e.g. `~/.sourcegraph`) does not yet exist. Once the
container is ran for the 2nd time, it starts as expected without any
intervention.

I've noted this in the update docs, and also explained it is possible to
pre-emptively create the directory with the correct permissions to avoid
running into the error at all, if desired.

Otherwise the bug is harmless, and the v4.3 release is due December 15th
so anyone worried about this can wait a few days for v4.3.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>