CodeMirror blob: implement "navigate to any line on click" (#45064)
* CodeMirror blob: implement "navigate to any line on click"

This commits makes it possible to promote any line from the ref panel
by clicking anywhere on that line in the preview pane. This feature is
supported by the old blob view, this commit just closes the feature gap.

This commit implements this functionality by wrapping all lines with an
anchor link. This is slightly different from the old blob view where
every line registers an ad-hoc click handler.

* Don't require props.nav to be defined

We can fallback to `history.push` if the custom action is not defined,
and this fallback could avoid a cryptic debugging session in the
future.

* Address review feedback and simplify code

I didn't know about `blobPropsFacet` when I implemented the original
commit. The code is much simpler when we can access the props via
`view.state.facet`.

* ESLint

* Remove unnecessary hook dependency