search: reduce cache length in case of error or rollout
If we encounter an error or we are rolling out, lets not cache that
information for as long. This caching behaviour only runs for ListAll
queries. I expect the biggest impact for this is during rollouts of
Zoekt we more quickly return to using Zoekt for all searches.

Test Plan: go test