search: handle backends missing when using repoHasContent
If zoekt is rolling out it was possible that between the List done in
PartitionRepos and the actual List call for reposHasContent that the
underlying index has changed. We can't exactly protect for this with our
current APIs. Instead, when a repoHasContent query is used during a
rollout, we add a BackendsMissing event to the search. For clients like
code-insights this will lead to retries/etc.

Note: I'm not a huge fan of needing to add the MaybeSendStats call. This
feels like we should adjust the design so we don't need each callsite to
do this. But just adding in stream to repos.Paginate feels wrong as
well. So this was the compromise that felt more explicit.

Test Plan: go test and CI