separate repo and rev before attempting to compile RepoNamePatterns regexps For the purpose of highlighting repo name matches, during RepoSearchJob construction we attempt to compile each value in RepoOptions.RepoFilters to regex and then populate RepoSearchJob.RepoNamePatterns with those compiled regexps. However, we weren't separating the revision from the overall repo filter pattern before doing this, so if the revision happened to be invalid regex then the code would panic.
This separates the repo and rev portions of each repo filter pattern so that we only try to compile the repo portion to regex. Also, it seems like when this code panicked it also prevented the "Invalid revision syntax" error from being surfaced. Now that error seems to be back as expected.