honey: prometheus counter around NewEvent calls
Main purpose of this commit is to identify services in production which
are attempting to log to honeycomb but do not have the secrets setup. In
particular I'm interested in ensuring all uses of the zoekt-search
dataset are getting logged.

Maybe we don't need a counter since we could just check counters in
honeycomb, but this felt natural.

Test Plan: go test