database: implement `ShareableStore` on stores that accept them
This was something @adeola-ak and I discovered: not all stores that are
part of the "shareable store ecosystem" (ie those that have `With`
methods) actually implement `basestore.ShareableStore` themselves.

This adds the missing methods to the interfaces — since all of the
concrete `struct`s embed `basestore.Store`, we don't actually have to
add any new implementations; we're just exposing their existing `Handle`
methods.