Always display search syntax help button (#45637)
Previously: the search syntax help button was not visible in the
`/search` page, which is the page where users are most likely to need
help with the syntax. This special-case got added in this commit
https://github.com/sourcegraph/sourcegraph/commit/3b1483197c12f88e8195347f98be770ec026e10d#diff-2fc185b787901924c9d5f03b8f28599b87896268a53070f3501afa42387e3d32R55
but without an explanation why the `/search` page is special cased.
Removing the special case makes the button appear again.