Feature Flags: reduce DB roundtrips (#45747)
Previously, we made three different DB roundtrips to calculate feature
flags for a user. One to get the list of defined feature flags, one to
get the list of user-specific overrides, and one to get the list of
org-specific overrides. We did this concurrently for performance.
However, doing this concurrently meant we frequently hit the "no
concurrent transactions" errors and log messages.

This updates our feature flag evaluation to use only a single database
query, joining the overrides with the defined flags in one fell swoop.