search: show indexserver on index status page (#45653)
It's a common problem to figure out on which indexserver a repo's index
resides. With this change we can finally look it up on the index status
page.

Note that this comes with the same limitation as the "Reindex now"
button: The host can only be resolved if the socket connection between
webserver and indexserver is working. In case it is not working, we
display "unknown: <error message>".