symbols: fix possible nil pointer dereference when GraphQL query contains no "query" clause in `symbols` resolver.
This is not guaranteed anywhere up the code stack (`query` is not a mandatory GraphQL parameter), that's why I added this check. I came across it during local testing with a query like:
```
query CommitSymbols($repoName: String!, $revision: String!) {
	repository(name: $repoName) {
		commit(rev: $revision) {
            symbols {
				nodes {
					name
				}
			}
		}
	}
}
```

Test plan:
Existing CI, because adding a unit test for `searchZoekt` looks non-trivial and the change is tiny.