show repo container UI even while a subcomponent is still loading
Previously, if the repo container was displaying a lazily-loaded component (for which a `fetch` was needed), the repo container header would not be displayed. This change makes it so that it is displayed. This improves the perceived loading time. It also fixes a bug where if a child component initially renders something and then shortly thereafter renders a suspended/lazily-loaded component, the entire repo container will blank out until that lazy-loaded component is loaded.