Search suggestions: complete on search values
Previously, typing `symbol` didn't suggest writing `type:symbol`.
Now, we additionally suggest all static filter values, including for the
filters `type`, `select`, `patterntype`, and `lang`. This means that we now
suggest `lang:typescript` when the user writes `typesc`, and we also
suggest `patterntype:regexp` if the user types `regexp`.

The diff is more involved that I was expecting because we want to avoid
the situation where static suggestions move from the top to the bottom
when dynamic suggestions appear. This required disabling the `filter`
option in favor of providing more detailed `boost` scores.

This commit also makes the matching algorithm more strict for static
values to avoid the situation where the user types `scip` and it matches
`TypeScript`.  To prevent that situation, the user must with an exact
substring of static values.