Revert "Revert "web: disable background feature flags evaluation (#45… (#45834)
Revert "Revert "web: disable background feature flags evaluation (#45647)" (#45813)"

E2E tests setup update: when we load the Sourcegraph web app, we need to wait for either one or two redirects to happen:

1. If the site init is not required, the user is redirected to the /sign-in URL.
2. If the site init is required, the user is redirected first to the /sign-in URL and then to the /site-admin/init URL.

We used to rely on the Puppeteer method to wait for all redirects (we waited for all network requests to complete). But this approach is not stable. With updates in this PR, we will first wait for the sign-in form selector to appear on the screen. If this operation fails with the timeout error, we wait for the sign-up form selector. If that fails, we fail the test suite.