remove code intel from secondary locations to ease extension API removal (#45920)
This removes code intel from marginal places to make it easier to fully remove the legacy extension API support.

Until we re-add it, these places will no longer have code intel (hovers):

- Batch Changes diff snippets
- Notebooks snippets
- Diffs (commits, comparisons)

Right now, these are implemented in a way that depends on the legacy extensions controller API. We want to remove that fully. That migration will go much, much faster if we don't need to maintain the legacy implementation for these secondary and lesser-used places.