webhooks: Factor our duplicate push handling code
We noticed that all the push handlers were doing the same thing:

1. Extracting the clone URL from an event
2. Getting the repo name from the clone URL
3. Queueing a repo update

The only thing specific to each code host is the part that extract the
clone URL so we add a generic method to do the rest.

Also, since the code shrunk so much it's now easier to read if we move
everything into a single file.