remove "click to go to definition" behavior
In the web app, the new selection-driven code nav UI (https://github.com/sourcegraph/sourcegraph/pull/44698) does not use this.

The browser extension's implementation is separate from the web app's in implementation and backstory, but it is also removed because it presents the same future UX problems and is inconsistent with the direction we're taking for the web app.