Search: Improve ref glob check for search context queries
When a search context is based on a query, we don't allow the repo filter to
include ref globs. The check for this was a bit fragile -- it just checked
whether RevisionSpecifier.RevSpec was empty, with the assumption that one of the
other glob fields would be set. But we also use an empty RevSpec to indicate the
default branch.

This commit makes the check more robust and adds a test. Hopefully the new
method RevisionSpecifier.IsRefGlob helps prevent this type of mistake.