fix apollo cache merge error for Repository by fetching ID in queries (#46063)
This fixes an Apollo GraphQL client error (shown in the devtools console):

`Cache data may be lost when replacing the [...] field of a Query object. To address this problem (which is not a bug in Apollo Client), either ensure all objects of type Repository have IDs, ...`

The simple and safe fix is to query the `Repository.id` field so that Apollo knows how to merge that data into its cache for a given Repository node.

There are probably more places where this should be done, but this fixes all the ones I encountered.