Insights: handle repo revs when filtering by search context (#46089)
Code insights allows filtering the repositories on-the-fly using a search
context. To support this, the insights code loads the search context query,
extracts its repo filter, and compiles the filter regex.

The search context repo filters can contain revisions too. The logic doesn't
account for this and just compiles the whole repo@revision string. This filter
will never match any repos. If the revision is not a valid regex, then it
can also fail to compile and cause a cryptic error.

Now we parse the repo filter and return an error when any revisions are
present. This lets the user know their context won't work as intended.