zoekt: default to eager runeDocSection (#46245)
We have been running with eager runDocSection decoding on
sourcegraph.com for a month and have seen improvements to tail latency
and average latency. We believe the lazy decoding was unnecessary since
we so often do global symbol searches that we would constantly be
decoding doc section data all the time.

There is a chance that on a quiet instance we suddenly have a lot more
RAM sitting in the heap that doesn't get claimed back since it stays
alive. In that case the lazy decoding may in fact be better for them. As
such we document how to disable in the CHANGELOG.

For more details see the PR in zoekt
https://github.com/sourcegraph/zoekt/pull/503

Test Plan: tested already on sourcegraph.com.