web: add support of created and updated dates without authors in CreatedByAndUpdatedByInfoByline.
Note: We cannot derive `noAuthor` from checking that both `createdBy` and `updatedBy` are null, because it clashes with a case when there is only created date and the user is deleted (we need to show "by a deleted user" in this case).

Test plan:
Storybook, local sg run and manual checks.