Add a note to fetch --all from reference repo. (#46337)
Currently, the docs say to add the upstream, but they don't ask the user to fetch branches and commits. This leads to the following error being returned whenever they attempt to create a release branch after adding the reference repository.

`fatal: 'v4.3.1' is not a commit and a branch 'release' cannot be created from it`

Adds a `git fetch --all` step to the flow.