prom-wrapper: remove direct Prometheus access, remove alerts endpoints (#46299)
This removes the one instance of Sourcegraph being able to directly query Prometheus. Likely not something we should have done in the first place, and will make it easier for Cloud the potentially run a completely-managed Prometheus. We retain some APIs for the embedded Alertmanager instance for configuration validation and the like - Alertmanager has a smaller surface area, and future alert summary queries should use the Alertmanager API instead.

The GraphQL endpoints have been marked as deprecated and no longer return any data. In the future, we could replace with this the endpoints that talk to Alertmanager to get alerts summaries - the Alertmanager API is a smaller surface area and we likely want to continue to self-host Alertmanager in the future for the ability to better configure alerting.

This endpoint is rarely used - the only two usages I could find are:

- report-bug page (awareness of this is very low, and the data is noisy - the integration has been removed in this change)
- instancehealth in src-cli - this is a non-critical, experimental integration and we can replace it with manual checks and/or a new API as described above