internal/database: Do not LIMIT 0 for default behaviour
A limit 0 does not really make sense from an API consumer's perspective
so it is expected that if no limit is set, the resulting DB query should
not LIMIT 0 by default. And if a limit is set in LimitOffset, then the
current behaviour is unchanged.

Let's see what breaks with this in the CI. 🤞🏽