search: Fix/improve client side query parser
The current implement incorrectly handles "leaves". For example 'a b OR c'
is parsed as

{
  type: 'operator',
  operands: [
    {type: 'pattern', value: 'a'},
    {type: 'pattern', value: 'b'},
    {type: 'pattern', value: 'c'},
  ]
}

which is wrong, or at least imprecise. Without more information it's not
clear whether this node represents 'a OR b c' or 'a b OR c'.

With this commit consecutive "leaves" are represented as a "sequence"
instead (see test output changes).