all: introduce redispool.KeyValue (#46425)
KeyValue is an abstraction over our most common uses of redis in our
codebase. It is introduced in an effort to have an alternative to redis
in Sourcegraph App. Additionally it provides a more ergonomic API than
the raw redis interface we have in our codebase.

The most important change is in keyvalue.go. Otherwise the changes to
all other files are much more minimal and should be backwards
compatible.

Note: Some call sites would do multiple commands on a single connection.
The only benefit here is performance, but in each case we did that it
was a function rarely called. Additionally in some places we used
redis.Conn.Send, but now we use redis.Conn.Do always. This means
previously if the command failed we wouldn't know since Send only places
the command into a buffer, and in all cases we ignored the error
returned from Close. Now we won't ignore that error.

Note: Currently Pool should always return, optionally disabling redis is
for a future commit.

Test Plan: unit tests, manual testing and will also do integration tests
in CI.