Search: Add metric for position of first click
This change adds a prometheus metric for the position of the user's first search
result click. Currently we just track the position for all clicks. It's helpful
to distinguish these two cases -- for example, if a user scrolls past the top
results and clicks on the 10th one, then there's a clear ranking problem. But if
they click on the 1st, 2nd, 3rd, then 10th results, this is actually pretty good.

It also adds a label for the number of search results. That way, we can exclude
cases when there was a tiny number of results (< 3) and the user clicked on one
of these. This doesn't indicate a ranking success.