temporary settings: Fix stale value issues with frequent updates
The temporary settings storage listens to changes from the backend and
keeps a local copy of the data. When the data is updated the update is
sent to the backend, the local copy is updated and subscribers are
notified of the update.
Sending an update to the backend will cause all subscribers to be
notified of the update after the backend processed it, including the
subscriber that made the update.

This can cause issues if multiple updates happen in quick succession,
something like:

```
subscriber        storage            backend
  |                  |                  |
  |                  |                  |
  | -- update A -->  |                  |
  | <-- notify A --  |                  |
  |                  | -- update A -->  | process A
  |                  |                  |
  | -- update B -->  |                  |
  | <-- notify B --  |                  |
  |                  | -- update B -->  | process A,B
  |                  |                  |
  |                  | <-- notify A --  | process B
  | <-- notify A --  |                  |             ! subscriber receives outdated data !
  |                  | <-- notify B --  |
  | <-- notify B --  |                  |
```

More specifically: The server backend uses Apollo's `watchQuery`
to notify the storage about any changes. It looks like it's possible that Apollo's
cache is update *after* another edit has already been made.

Interestingly I couldn't observe affecting this usage in React, even
though I can see that the data returned by `useTemporarySetting` is
stale. But I noticed this in the implementation I've worked on for the
SvelteKit prototype.