frontend: Fix subtle bug in observeSystemIsLightTheme (#46709)
frontend: Fix subtle bug in observeSystemIsLight

At the moment is possible to that the observable emits incorrect data to
new subscribers. I assume the expectation is that the function is called
every time a callsite wants to subscribe, or that the observable is only
used once but that's not obvious.

If the observable is indeed reused it would emit the value of
`!mediaList.matches` that it had at the time the function was called.
They user might have changed their theme preference in the meantime.

By using `defer` we evaluate the value of `mediaList.matches` every time
a subscriber subscribes.