rcache: add TTL, Expire and Incr to KeyValue (#46780)
Was hoping we could nicely avoid making the KeyValue interface even
larger, but no dice. Lets do this and see how we can make it not such a
large thing. Additionally the test for KeyValue has become quite large,
but I believe is still readable and debuggable. But there is scope to
improve that in the future as well.

Test Plan: added unit tests. Existing ones test rcache methods changed
as well.