redispool: implement in memory KeyValue via minimal interface
This implementation is targetted at making it very simple to implement a
redis like store by only implementing a key value store which stores a
byte slice for a key.

Right now it is unused, but has extensive test coverage as well. It will
be used in the Sourcegraph App in the next commit.

Test Plan: go test