Search: fix validation for repo filters (#46801)
We always validate queries before executing them to ensure each field has the
right type, there aren't conflicts between filters, etc. The validation for repo
filters was slightly incorrect, because it tested if the whole repo filter could
be parsed as a regexp. Instead we should split the filter into repo and revs,
and compile the repo part. This is because in rare cases like "[@rev]", the
whole filter is a valid regexp, but the repo part is not.

We weren't catching this case before, which meant downstream code could
encounter invalid repo filters and panic.