executors: always use in-memory password for Sourcegraph App
This change makes us actually always use the in-memory password when the
deployment type is Sourcegraph App.

After #46721 the executor password was randomly generated, but the site
config had the randomly generated password written once, while the env
var would always use the latest in-memory password. Effectively, the one
saved to disk could diverge from the one in memory on the second run.

In this change I instead have us respect the in-memory password at all
locations that we use the configured password, iff the deployment type
is App.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>