Fix error handling when github oauth fails to create a user with signups allowed
Previously, when github oauth provider had `allowSignup: true` configured,
we returned the first error. However the code is written in a way, that it
first tries to match existing user accounts to a list of verified email addresses
from github. It only tries the signup flow when there is no match.

If for any reason the signup flow fails, e.g. a license is breached, the user
got the first error saying there is no match on the email addresss, which is bogus.

Fixing this problem by returning the last error instead. Also added a regression
unit test, so that this does not happen in the future.