experimental query input: Add support for file completions and suggestions (#46663)
Co-authored-by: Juliana Peña <me@julip.co>

This is based on the recent context filter and query parser work.

For dotcom:

- File suggestions are only shown when the query contains at least one context: filter different from global or a repo: filter.
- file: value completions work whether the query contains context: and repo: filters or not.

For customer instances:

- File suggestions are always shown whether the query contains context: and repo: filters or not.
- file: value completions work the same as for dotcom.

In both cases suggestions and completions are scoped to the relevant context: and repo: filters if present. That means for a given position in the query we traverse up the tree and keep all repo: and context: filters of the (explicit or implicit) AND branches.

I also updated repo: suggestions and completions to take into account context: selectors.