Disable connection checks for code hosts
This is currently implemetned as TCP dial to the code host. We have customers
behing a proxy, that disallows TCP dial, which means this check will always fail
for them, which is a false positive. On top of that, we do this check in repo
syncer as well. If the check fails, we disable repo syncing.

We decided a safer approach for now is to disable the feature and create a patch
release, so that we can unblock the affected customers.

We will fix these checks in a followup PR.