Change code host connection checks to talk HTTP (#46918)
Full context: https://github.com/sourcegraph/sourcegraph/issues/46915

This changes the connection checks to test what we're after: can we talk
to the code hosts' API via HTTP?

Because as it turns out, in some customer environments we can't do a TCP
dial, because proxies block it. In that case an HTTP call makes even
more sense, because if that succeeds we know we can talk to the API.

The other change in here is to not hard-fail the repository syncing if
the connection check fails. Until we're confident that the checks work
exactly as we want them to, we only print a warning.