gRPC: add panic handling interceptors (#46991)
Most of our HTTP APIs add a top-level panic handler so that we don't crash the process when a request panics. Symbols is no exception. However, when we switched to the gRPC implementation, that wasn't implemented, so an existing panic caused symbols to start crashing.

This adds an interceptor to the default set of gRPC options that logs the error and makes the RPC return an internal error with some details.