Search: restore old syntax for repo:contains filters (#47007)
In 4.0, we renamed the `repo.contains` predicates to be more consistent. At the
time, we didn't think any users would be affected, so we cut over fully to the
new syntax. But some users have scripts or batch changes that use the old
syntax, which will break on 4.0 with no grace period to cut over.

This change restores the old syntax alongside the new, to give users a time
period to transition. We still plan to remove the old syntax in the future. The
old syntax is only restored in the backend, it's not documented or part of
autocomplete.