Propose teams GraphQL/DB schema and add a first cut of docs (#45817)
This PR suggests an API schema, and how to use it via a small docs page, including a src-cli extension.

We will add working examples for how to ingest data from GitHub and GitLab for reference implementations of customer internal systems.

This aims to follow what has been described in RFC 772. Note that this PR doesn't yet address metadata on teams, this will be introduced in a separate PR.

Teams will not be an enterprise extension, given how deeply integrated they have to become. Syncers and custom integration later on might be though.