set meta noindex on doc pages that are not for the current version
Fixes a bug where noindex was not being set correctly on doc pages on non-current versions. This bug was introduced when new code was added to the template to also noindex pages related to Code Insights for the early 2021 launch. That embargo is no longer needed, so we can remove that code altogether.

After this change, visiting a page with an `@version` in the URL such as https://docs.sourcegraph.com/@3.1/dev/roadmap and inspecting the source should reveal a meta noindex tag.