Search: remove incorrect adjustment to zoekt MaxWallTime (#47058)
There's some old logic that tries to adjust SearchOptions.MaxWallTime when the
user sets a larger result limit, in order to give zoekt more time to find
matches. But for streaming search, this always fires, since we check against the
wrong default limit (we use the non-streaming limit). So we were always
adjusting the MaxWallTime to 50 seconds.

This PR removes the adjustment, since it's often broken and doesn't seem
helpful.