Ranking: add an experimental setting for tuning DocumentsRankWeight
This PR adds an experimental feature setting that allows for adjusting
DocumentRanksWeight. This will help us tune the contribution of the file ranks in
the overall search ranking, by playing around with this setting on test data.

It also fixes a bug where we accidentally disable a repo search optimization
when file-based ranking is turned on.