redispool: use in memory KeyValue for Sourcegraph App
We introduce a new function NewKeyValue which takes an address to dial
redis. We then check if the address is instead referring to in memory
and return an in memory key value store.

How we decide on address also had to change. We now ensure the address
is set before anyone can read the address variable. Previously we had
the addr accessed inside of a closure, so we didn't need that gaurentee.
We then changed the logic to use in memory as the default if the deploy
type is Sourcegraph App.

I expiremented with a logger to log every interaction we have with the
store. On startup it generates about 3mb of logs, not so bad.

Note: this uses in memory for Store. A follow up commit will switch
Store to using something persistent.

Test Plan: I ran the app with the follow command and ensured I didn't
have any redis envvars set in my environment. Additionally I added
logging to ensure the in memory store was used. I then did a few
searches and played around a bit.

  export USE_EMBEDDED_POSTGRESQL=0
  export DEPLOY_TYPE=single-program
  export SRC_FRONTEND_INTERNAL=localhost:3090
  export EXECUTOR_FRONTEND_PASSWORD=asdf
  export SRC_DISABLE_OOBMIGRATION_VALIDATION=1
  export BLOBSTORE_DATA_DIR=/tmp/blobstore-data
  go run ./enterprise/cmd/sourcegraph/