Search subexpressions proofread
Lines 20, 31, 64: Removed period to match formatting in "Useful search examples".
Line 53: Removed comma from "and e.g. share with coworkers."
Line 83: Moved paragraph to a single line. Changed "the" to "this" in "there are multiple ways to group this query".
Line 112: Moved paragraph to a single line. Removed comma after "is" in "what the intent is so that".